# students/k3340/laboratory_works/Bogdanov_Maxim/Lr3/docs/swagger/swagger.yaml
basePath: /api/v1
definitions:
...
  internal_delivery_http_handler.ClassResponse:
    properties:
      academic_year_id:
        type: integer
      class_teacher_id:
        type: integer
      created_at:
        type: string
      grade:
        type: integer
      id:
        type: integer
      letter:
        type: string
      updated_at:
        type: string
    type: object
  internal_delivery_http_handler.CreateClassRequest:
    properties:
      academic_year_id:
        type: integer
      class_teacher_id:
        type: integer
      grade:
        type: integer
      letter:
        type: string
    type: object
...
  internal_delivery_http_handler.UpdateClassRequest:
    properties:
      academic_year_id:
        type: integer
      class_teacher_id:
        type: integer
      grade:
        type: integer
      letter:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@school-service.com
    name: API Support
  description: API для управления школьной системой
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: School Management Service API
  version: "1.0"
paths:
...
  /classes:
    get:
      consumes:
      - application/json
      description: Возвращает список всех классов
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_delivery_http_handler.ClassResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Список классов
      tags:
      - classes
    post:
      consumes:
      - application/json
      description: Создает новый класс в системе
      parameters:
      - description: Данные класса
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_delivery_http_handler.CreateClassRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_delivery_http_handler.ClassResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Создать класс
      tags:
      - classes
  /classes/{id}:
    delete:
      consumes:
      - application/json
      description: Выполняет soft delete класса
      parameters:
      - description: ID класса
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Удалить класс
      tags:
      - classes
    get:
      consumes:
      - application/json
      description: Возвращает информацию о классе по его ID
      parameters:
      - description: ID класса
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_delivery_http_handler.ClassResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Получить класс по ID
      tags:
      - classes
    put:
      consumes:
      - application/json
      description: Обновляет информацию о классе
      parameters:
      - description: ID класса
        in: path
        name: id
        required: true
        type: integer
      - description: Данные класса
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_delivery_http_handler.UpdateClassRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_delivery_http_handler.ClassResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Обновить класс
      tags:
      - classes
  /classes/academic-year/{academicYearId}:
    get:
      consumes:
      - application/json
      description: Возвращает список всех классов указанного учебного года
      parameters:
      - description: ID учебного года
        in: path
        name: academicYearId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_delivery_http_handler.ClassResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Список классов учебного года
      tags:
      - classes
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"